refactor to move the clients together
