+++ packages/node/src/index.ts
@@ -1,5 +1,5 @@
 export { Event, Options, Response, Status } from '@amplitude/types';
-export { NodeClient } from './nodeClient';
+export { NodeClient, SimpleClient } from './clients';
 export { RetryHandler } from './retryHandler';
 export { init } from './sdk';
 export { HTTPTransport } from './transports';
